requirements.txt
apache-airflow==3.0.1
apache-airflow-providers-docker
requests
psycopg2-binary
pandas
sqlalchemy
dbt-core==1.9.0
dbt-postgres==1.9.0